processor/requirements.txt
python-dotenv
boto3
pika
msgpack
opencv-python
scipy
numpy
gdown